fix: wait for ads1299 reset before reading
In the current implementation, sending a reset request results in a
JSON decoding error. The HackEEG Arduino driver has a delay after reset,
before sending the response, which is not accounted for in the python
client.

This commit adds an optional delay between sending commands and
reading responses. The reset command is then configured with a 2s delay,
which is longer than the delay in the Arduino driver.